Extract reusable analysis engine for #263 and #264
Split the pure document-reduction and result-finalization logic into core/engine.js while keeping the shell-facing behavior unchanged.

Refs #263 and #264.
