+++ CONTRIBUTING.md
@@ -14,21 +14,25 @@ As an additional (not required) dependency, [Docker](https://www.docker.com/) or
 
 ## Repo Layout and the `variety.js` Build
 
-`variety.js` at the repo root is a generated file, assembled from four sources in this order:
+`variety.js` at the repo root is a generated file, assembled from five sources in this order:
 
 - `core/formatters/ascii.js` — built-in ASCII table formatter. Self-contained
   IIFE; registers an `ascii` factory on `shellContext.__varietyFormatters`.
 - `core/formatters/json.js` — built-in JSON formatter. Self-contained IIFE;
   registers a `json` factory on `shellContext.__varietyFormatters`.
-- `core/analyzer.js` — pure, transport-agnostic analysis logic. No runtime
-  dependencies on shell globals, Node I/O, or any other layer. Reads
-  `shellContext.__varietyFormatters` to dispatch output. This is the
-  future `@variety/core` package boundary.
+- `core/engine.js` — pure, side-effect-free analysis logic. No runtime
+  dependencies on shell globals, Node I/O, persistence, or formatting. Exports
+  the reducer/finalization engine used by both the shell-facing analyzer and
+  Node-side tests. This is the future `@variety/core` package boundary.
+- `core/analyzer.js` — shell-adjacent orchestration that traverses the cursor,
+  optionally persists results, dispatches to formatters, and exposes `run()`
+  on top of `core/engine.js`. Depends on `core/formatters/` and `core/engine.js`.
 - `mongo-shell/adapter.js` — the shell-facing layer that reads shell
   globals (`collection`, `plugins`, `secondaryOk`, etc.), loads plugins, and
   hands dependencies to `impl.run()`. The only place in the build that
-  touches `db`, `print`, and `load`. Cleans up both `__varietyImpl` and
-  `__varietyFormatters` after execution so repeated loads are idempotent.
+  touches `db`, `print`, and `load`. Cleans up `__varietyEngine`,
+  `__varietyImpl`, and `__varietyFormatters` after execution so repeated loads
+  are idempotent.
   Depends on `core`; compiled into `variety.js` by `build.js` (not a
   separately published package).
 - `bin/variety` — the published Node entrypoint that implements the main
@@ -42,10 +46,12 @@ As an additional (not required) dependency, [Docker](https://www.docker.com/) or
   `@variety/cli` package boundary.
 
 **Dependency directions:** `core/formatters/` has no runtime deps on any other
-layer. `core` depends on `core/formatters/`. `mongo-shell/adapter.js` depends on
-`core`. `mongo-shell/launcher.js` is Node-only (no `core` dep). `cli` depends on
-`core` and `mongo-shell/`. `build.js` composes `core/formatters/` + `core` +
-`mongo-shell/adapter.js` → `variety.js`.
+layer. `core/engine.js` has no runtime deps on any other layer. `core/analyzer.js`
+depends on `core/engine.js` plus `core/formatters/`. `mongo-shell/adapter.js`
+depends on `core/analyzer.js`. `mongo-shell/launcher.js` is Node-only (no `core`
+dep). `cli` depends on `mongo-shell/` and the built script path, not on the
+engine directly. `build.js` composes `core/formatters/` + `core/engine.js` +
+`core/analyzer.js` + `mongo-shell/adapter.js` → `variety.js`.
 
 `build.js` concatenates those source files under a generated-file banner.
 Edit the sources in `core/` or `mongo-shell/adapter.js`, then run:
