+++ build.js
@@ -3,8 +3,8 @@
 // SPDX-FileCopyrightText: © 2026 James Cropcho <numerate_penniless652@dralias.com>
 'use strict';
 
-// Assembles variety.js by concatenating the formatter files, core/analyzer.js, and
-// mongo-shell/adapter.js underneath a generated-file banner. The built variety.js
+// Assembles variety.js by concatenating the formatter files, core/engine.js,
+// core/analyzer.js, and mongo-shell/adapter.js underneath a generated-file banner. The built variety.js
 // is committed to the repository so `mongosh variety.js` works from a fresh clone without
 // running a build step first; CI runs this script and fails if the
 // committed file drifts from its sources.
@@ -28,7 +28,7 @@ Please see https://github.com/variety/variety for details. */
 //
 // Assembled by build.js from:
 //   core/formatters/ascii.js, core/formatters/json.js,
-//   core/analyzer.js, mongo-shell/adapter.js.
+//   core/engine.js, core/analyzer.js, mongo-shell/adapter.js.
 // To change behavior, edit those source files and run \`npm run build\`. The
 // build output is committed so \`mongosh variety.js\` works from a fresh clone
 // without a build step; CI verifies the committed file matches its sources.
@@ -44,20 +44,25 @@ Please see https://github.com/variety/variety for details. */
 // See .eslint.config.js for the enforced rule set.
 
 // -----------------------------------------------------------------------------
-// This file is organized in four sections, sourced from four separate files:
+// This file is organized in four sections, sourced from five separate files:
 //
 //   1. FORMATTER SECTION (core/formatters/ascii.js, core/formatters/json.js) —
 //      built-in output formatters. Each is a self-contained IIFE that registers
 //      a factory function on \`shellContext.__varietyFormatters\`. Third-party
 //      formatters can be supplied as plugins instead (see README).
 //
-//   2. IMPLEMENTATION SECTION (core/analyzer.js) — pure, transport-agnostic
-//      analysis logic. Functions take their dependencies (config, and where
-//      needed a \`log\` function or a \`deps\` bag holding shell primitives)
-//      as explicit parameters. The section hands a bundle of functions to
-//      the interface section via \`shellContext.__varietyImpl\`.
+//   2. ENGINE SECTION (core/engine.js) — pure, side-effect-free analysis logic.
+//      Functions take their dependencies (config, and where needed a \`log\`
+//      function) as explicit parameters and return structured analysis rows.
+//      The section hands a reusable engine to later sections via
+//      \`shellContext.__varietyEngine\`.
 //
-//   3. INTERFACE SECTION (mongo-shell/adapter.js) — everything that touches
+//   3. ANALYZER SECTION (core/analyzer.js) — shell-adjacent orchestration for
+//      cursor traversal, optional persistence, and formatter dispatch. Depends
+//      on the engine and hands the combined internal API to the interface
+//      section via \`shellContext.__varietyImpl\`.
+//
+//   4. INTERFACE SECTION (mongo-shell/adapter.js) — everything that touches
 //      shell globals: reading input (\`collection\`, \`plugins\`, \`__quiet\`,
 //      \`secondaryOk\`, etc.), the config-echo logging, plugin loading via
 //      \`load()\`, input validation, and constructing the dependency bag
@@ -71,12 +76,13 @@ Please see https://github.com/variety/variety for details. */
 const root = __dirname;
 const fmtAscii = fs.readFileSync(path.join(root, 'core', 'formatters', 'ascii.js'), 'utf8');
 const fmtJson  = fs.readFileSync(path.join(root, 'core', 'formatters', 'json.js'), 'utf8');
-const impl     = fs.readFileSync(path.join(root, 'core', 'analyzer.js'), 'utf8');
+const engine   = fs.readFileSync(path.join(root, 'core', 'engine.js'), 'utf8');
+const analyzer = fs.readFileSync(path.join(root, 'core', 'analyzer.js'), 'utf8');
 const iface    = fs.readFileSync(path.join(root, 'mongo-shell', 'adapter.js'), 'utf8');
 
 // HEADER ends with a single \n; each source file ends with a single \n. We
 // want two blank lines (three \n total) between each region in the output.
-const output = `${HEADER}\n\n${fmtAscii}\n\n${fmtJson}\n\n${impl}\n\n${iface}`;
+const output = `${HEADER}\n\n${fmtAscii}\n\n${fmtJson}\n\n${engine}\n\n${analyzer}\n\n${iface}`;
 
 const outPath = path.join(root, 'variety.js');
 
@@ -85,7 +91,7 @@ if (args.includes('--check')) {
   const existing = fs.readFileSync(outPath, 'utf8');
   if (existing !== output) {
     process.stderr.write(
-      'variety.js is out of date relative to its sources (core/formatters/, core/analyzer.js, mongo-shell/adapter.js).\n' +
+      'variety.js is out of date relative to its sources (core/formatters/, core/engine.js, core/analyzer.js, mongo-shell/adapter.js).\n' +
       'Run `npm run build` and commit the updated variety.js.\n'
     );
     process.exit(1);
