+++ test/cases/analysis/DeprecatedBsonTypesTest.js
@@ -6,8 +6,7 @@ import { fileURLToPath } from 'url';
 import { BSONSymbol, Code, DBRef, ObjectId } from 'mongodb';
 
 const require = createRequire(import.meta.url);
-const analyzerPath = fileURLToPath(new URL('../../../core/analyzer.js', import.meta.url));
-require(analyzerPath);
+const enginePath = fileURLToPath(new URL('../../../core/engine.js', import.meta.url));
 
 /**
  * @typedef {{
@@ -20,15 +19,12 @@ require(analyzerPath);
  *   maxExamples: number,
  * }} AnalyzerConfig
  * @typedef {{
- *   varietyTypeOf: (config: AnalyzerConfig, thing: unknown) => string
+   *   varietyTypeOf: (config: AnalyzerConfig, thing: unknown) => string
  * }} VarietyImpl
  */
 
-const analyzerGlobal = /** @type {typeof globalThis & { __varietyImpl?: VarietyImpl }} */ (globalThis);
-const impl = analyzerGlobal.__varietyImpl;
-if (typeof impl === 'undefined') {
-  throw new Error('Expected core/analyzer.js to register __varietyImpl.');
-}
+const rawImpl = /** @type {unknown} */ (require(enginePath));
+const impl = /** @type {VarietyImpl} */ (rawImpl);
 
 /** @type {AnalyzerConfig} */
 const config = {
